+++ .types.lua
@@ -51,6 +51,7 @@ TSM_API = {}
 ---@class ProfessionTransaction
 ---@field allocationTbls ProfessionTransationAllocations[]
 ---@field GetRecraftAllocation fun(self: self): string
+---@field GetRecipeSchematic fun(self: self): CraftingRecipeSchematic
 ---@field CreateCraftingReagentInfoTblIf fun(self: self, predicate: function): CraftingReagentInfo[]
 ---@field CreateCraftingReagentInfoTbl fun(self: self): CraftingReagentInfo[]
 ---@field CreateOptionalOrFinishingCraftingReagentInfoTbl fun(self: self): CraftingReagentInfo[]